requirements.txt
pip==23.3
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.22
django-cors-headers==4.1.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environ==1.0
gunicorn==21.2.0
inflection==0.5.1
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.7
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
